add ember liquid fire modal example
This is a basic example of liquid fire modal windows.

One tweak is that this project has a 'session' service object that is injected into routes, controllers and components.

Injecting into the component feels kind of strange but is an easy way to make the session object available to the modal component.